little code refactor

+++ app/src/main/java/pl/bsk/chatapp/model/FileMeta.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 class FileMeta(
     val filename: String,
-    val size: Long,
+    val size: Int,
 ): Serializable {
     override fun toString(): String {
         return "FileMeta(filename='$filename', size=$size)"
